+++ plugins/main/public/components/common/dashboards/panel.tsx
@@ -21,8 +21,8 @@ import {
 import { CustomSearchBar } from '../custom-search-bar';
 import { LoadingSearchbarProgress } from '../loading-searchbar-progress/loading-searchbar-progress';
 import { FILTER_OPERATOR } from '../data-source';
-import { WAZUH_SAMPLE_ALERTS_CATEGORY_SECURITY } from '../../../../common/constants';
 import { compose } from 'redux';
+import { PatternDataSourceFilterManager as dataSourceFilterManager } from '../../common/data-source/pattern/pattern-data-source-filter-manager';
 
 /**
  * Create a panel dashboard component using minimal dependencies: datasource, sample data warning
@@ -78,7 +78,6 @@ export const createPanel = ({
         isLoading: isDataSourceLoading,
         fetchData,
         setFilters,
-        filterManager,
         searchBarProps,
       },
     }) => {
@@ -86,8 +85,9 @@ export const createPanel = ({
         field: '',
         value: '',
       });
-      const [currentSelectedFilter, setCurrentSelectedFilter] = useState();
-      const [selectedPanelFilter, setSelectedPanelFilter] = useState([]);
+      const [selectedPanelFilter, setSelectedPanelFilter] = useState<
+        { [key: string]: any }[]
+      >([]);
       const filterDrillDownValue = value => {
         setDrillDownValue(value);
       };
@@ -100,18 +100,18 @@ export const createPanel = ({
         const { field, value } = selectedFilter;
         const controlledByFilter = 'office-panel-row-filter';
         if (value) {
-          const filter = filterManager?.createFilter(
-            FILTER_OPERATOR.IS_ONE_OF,
-            field,
-            [value],
-            controlledByFilter,
-          );
+          const filter: { [key: string]: any } =
+            dataSourceFilterManager?.createFilter(
+              FILTER_OPERATOR.IS_ONE_OF,
+              field,
+              [value],
+              controlledByFilter,
+            );
           // this hide the remove filter button in the filter bar
           setSelectedPanelFilter([filter]);
         } else {
           setSelectedPanelFilter([]);
         }
-        setCurrentSelectedFilter(selectedFilter);
       };
 
       return (