+++ CHANGELOG.md
@@ -18,6 +18,7 @@ All notable changes to the Wazuh app project will be documented in this file.
 
 - Updated OS logos [#7712](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7712)
 - Changed "Services" tab label to "Listeners" in IT Hygiene > Networks [#7742](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7742)
+- Removed "Top 5 agents" visualization and changed "Top 5 groups" label to "Groups" in IT Hygiene > Identity > Groups [#7792](https://github.com/wazuh/wazuh-dashboard-plugins/pull/7792)
 
 ### Fixed